requirements.txt
PyQt5
PyMuPDF
PyPDF2
pytesseract
Pillow
